Refactor: Node.java - changed datatype of x,y to int

+++ src/utils/Node.java
@@ -2,26 +2,26 @@ package utils;
 
 public class Node {
 	
-	private float x,y;
+	private int x,y;
 	
-	public Node(float x, float y) {
+	public Node(int x, int y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public float getX() {
+	public int getX() {
 		return x;
 	}
 
-	public void setX(float x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 
-	public float getY() {
+	public int getY() {
 		return y;
 	}
 
-	public void setY(float y) {
+	public void setY(int y) {
 		this.y = y;
 	}
 }
